requirements.txt
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.2.1
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.0
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
moviepy==2.1.2
mpmath==1.3.0
networkx==3.3
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==10.4.0
proglog==0.1.10
pyav==14.2.1
pycparser==2.22
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==75.8.0
soundfile==0.13.1
sympy==1.13.1
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
